requirements_github_action.txt
dbt-duckdb==1.5.2
dlt==0.3.8
google-api-python-client==2.84.0
google-cloud-bigquery==3.10.0
google-cloud-storage==2.8.0
msgpack==1.0.5
pyarrow==12.0.1